+++ MultiTrain/utils/__init__.py
@@ -1,3 +1,1 @@
-"""Compatibility exports for MultiTrain's shared utility helpers."""
-
 from MultiTrain.utils.utils import *